Rearrange all other recipes in their items' files
