+++ Content/Items/UpgradeKitCorruption.cs
@@ -2,6 +2,7 @@
 using BiomeExtractorsMod.Content.Tiles;
 using Terraria;
 using Terraria.Enums;
+using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace BiomeExtractorsMod.Content.Items
@@ -15,6 +16,15 @@ namespace BiomeExtractorsMod.Content.Items
         {
             base.SetDefaults();
             Item.SetShopValues(ItemRarityColor.Green2, Item.buyPrice(gold: 5)); // sell at 1
+        }
+
+        public override void AddRecipes()
+        {
+            CreateRecipe()
+                .AddIngredient(ItemID.DemoniteBar, 5)
+                .AddIngredient(ItemID.VilePowder, 12)
+                .AddTile(TileID.TinkerersWorkbench)
+                .Register();
         }
     }
 }
