+++ Content/Items/UpgradeKitCyber.cs
@@ -2,6 +2,7 @@
 using BiomeExtractorsMod.Content.Tiles;
 using Terraria;
 using Terraria.Enums;
+using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace BiomeExtractorsMod.Content.Items
@@ -16,6 +17,15 @@ namespace BiomeExtractorsMod.Content.Items
             Item.width = 18;
             Item.height = 18;
             Item.SetShopValues(ItemRarityColor.Lime7, Item.buyPrice(gold: 5)); // sell at 1
+        }
+
+        public override void AddRecipes()
+        {
+            CreateRecipe()
+                .AddIngredient(ItemID.ChlorophyteBar, 5)
+                .AddIngredient(ItemID.Nanites, 12)
+                .AddTile(TileID.MythrilAnvil)
+                .Register();
         }
     }
 }
